+++ eras/shelley/impl/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.15.0.0
 
-* Remove export of deprecated `RewardAcnt`
+* Remove export of deprecated `RewardAcnt` and `poolCWitness`
 * Change type of VRF key hash in `GenesisDelegCert`, `GenesisDelegTxCert` and `DuplicateGenesisVRFDELEG` to `VRFVerKeyHash`
 * Added `EncCBOR` instance for `LedgerEnv`
 * Use `Mismatch` to clarify _some more_ predicate failures. #4711